zach's comments
